chore(legacy): archive capture helpers and clean agent runtime
