+++ app/_legacy/capture/__init__.py
@@ -1,0 +1,3 @@
+"""
+Legacy capture pipeline utilities retained for reference.
+"""
