+++ app/services/policy.py
@@ -1,11 +1,5 @@
-import hashlib
-from typing import Dict
+"""
+Compat shim pointing to legacy capture policy helpers.
+"""
 
-def invariants_ok(text:str)->bool:
-    return True
-
-def compute_hashes(text:str)->Dict[str,str]:
-    return {"content_hash": hashlib.sha256(text.encode("utf-8")).hexdigest()}
-
-def version_vector(a:str,b:str):
-    return []
+from app._legacy.services.policy import *  # noqa: F401,F403
