+++ tests/legacy/__init__.py
@@ -1,0 +1,1 @@
+"""Legacy test placeholders retained for historical reference."""
