+++ tests/legacy/test_capture_ingest_roundtrip.py
@@ -1,0 +1,7 @@
+import pytest
+
+pytest.skip(
+    "legacy pre-v4.4 capture_ingest contract; CLI is now ObjectStore-backed and "
+    "emits Outbox events atomically. This test is kept for historical reference.",
+    allow_module_level=True,
+)
