+++ tests/legacy/test_vault_sync.py
@@ -1,0 +1,7 @@
+import pytest
+
+pytest.skip(
+    "legacy vault_sync bootstrap flow retained for reference. "
+    "Promotion lifecycle tests supersede this scenario.",
+    allow_module_level=True,
+)
